Update settings.py
When the envfile line "DEBUG=0" is parsed, the 0 is a string. However, bool('0') returns True since '0' is not empty. This means that settings.DEBUG is True as a result of the envfile line "DEBUG=0", which seems incorrect.  Casting the result of the get() call with int() fixes this, since bool(0) is False.